AMLS-4988: Splits ConfirmationController into 4
