+++ app/views/payments/bank_details.scala.html
@@ -96,6 +96,6 @@
         gaTag=Some("bank-details:click:print")
     )
 
-    <p><a class="button" id="banks-details-submit" href="@controllers.routes.ConfirmationController.bacsConfirmation()">@Messages("button.continue")</a></p>
+    <p><a class="button" id="banks-details-submit" href="@controllers.routes.BacsConfirmationController.bacsConfirmation()">@Messages("button.continue")</a></p>
 
 }
